+++ app/src/main/java/com/huy/fitsu/data/model/Budget.kt
@@ -9,8 +9,8 @@ import java.util.*
 @Entity(tableName = "budgets")
 data class Budget(
     @PrimaryKey @ColumnInfo(name = "id") val id : String = UUID.randomUUID().toString(),
-    @ColumnInfo(name = "value") val value : Int = 0,
-    @ColumnInfo(name = "expense") var expense : Int = 0,
+    @ColumnInfo(name = "value") val value : Float = 0F,
+    @ColumnInfo(name = "expense") var expense : Float = 0F,
     @Embedded val semanticWeek: SemanticWeek,
     @ColumnInfo(name = "budget_duration") val budgetDuration: BudgetDuration = BudgetDuration.WEEKLY
 )
